// moderateur/src/app/models/reservation.ts
export class Reservation {
    id?: string;
  reference?: string;
  etat?: string;
  typeCharge?: string;
  typeConnecteur?: string;
  dateReservation?: string;
  dateActuelle?: string;
  period?: string;
  createdAt?: string;
  updatedAt?: string;
  BornConnectId?: string;
  ClientId?: string;
  Client_id?: string;
  firstname?: string;
  lastname?: string;
  login?: string;
  password?: string;
  email?: string;
  image?: string;
  phone?: string;
  loginType?: string;
  codeVerif?: string ;
  etatVerif?: string;
  connexion?: string ;
  BlackList?: string;
  Client_createdAt?: string;
  Client_updatedAt?: string;
  BornConnect_id?: string;
  BornConnect_etat?: string;
  BornConnect_createdAt?: string;
  BornConnect_updatedAt?: string;
  BornConnect_BornId?: string;
  BornConnect_ConnecteurId?: string;
  Born_id?: string;
  Born_serialNumber?: string;
  Born_etatB?: string;
  Born_proprietaire?: string;
  Born_pourcentageP?: string;
  Born_NumeroTel?: string;
  Born_name?: string;
  Born_createdAt?: string;
  Born_updatedAt?: string;
  Born_StationId?: string;
  Born_Station_id?: string;
  Born_Station_registreNumber?: string;
  Born_Station_adresse?: string;
  Born_Station_name?: string;
  Born_Station_etat?: string;
  Born_Station_stationType?: string;
  Born_Station_confirmationAdmin?: string;
  Born_Station_prixKW?: string;
  Born_Station_telf?: string;
  Born_Station_longitude?: string;
  Born_Station_latitude?: string;
  Born_Station_createdAt?: string;
  Born_Station_updatedAt?: string;
  Born_Station_AdminId?: string;
  Born_Station_ProprietaireId?: string;
}